refactor: 엔티티 ai 서치 구조 변경 (#78)
