+++ src/app/routes/index.tsx
@@ -19,7 +19,7 @@ import UserComments from "@/pages/user/userpage/UserComments.tsx";
 import UserFollowers from "@/pages/user/userpage/UserFollowers.tsx";
 import UserFollowees from "@/pages/user/userpage/UserFollowees.tsx";
 import MyPage from "@/pages/user/MyPage";
-import WebtoonSearchResults from "@/pages/search/WebtoonSearchResults.tsx";
+import WebtoonSearchResults from "@/pages/webtoon-search/WebtoonSearchResults.tsx";
 import Error from "@/pages/error/Error.tsx";
 import MyProfile from "@/pages/user/mypage/MyProfile";
 import AdminMain from "@/pages/admin/AdminMain";
@@ -34,7 +34,7 @@ import Stats from "@/pages/admin/stats/Stats";
 import FeedbackStatus from "@/pages/admin/recommendation/FeedbackStatus";
 import UserAnalysis from "@/pages/admin/visualization/UserAnalysis";
 import TagManagement from "@/pages/admin/settings/TagManagement";
-import TextBasedRecommendations from "@/pages/text-based-recommendations/TextBasedRecommendations.tsx";
+import TextBasedRecommendations from "@/pages/webtoon-search-ai/TextBasedRecommendations.tsx";
 import AIRecommendation from "@/pages/AIRecommendation/AIRecommendation";
 import IdResult from "@/pages/auth/find/IdResult.tsx";
 import FindIdSecurityQuestion from "@/pages/auth/find/FindIdSecurityQuestion";
