--- src/entities/navigation/model/types.ts
@@ -1,10 +1,0 @@
-export interface NavItem {
-    title: string
-    href: string
-    value: string
-}
-
-export interface SubTabItem {
-    title: string
-    href: string
-}
