--- src/entities/notification/model/types.ts
@@ -1,4 +1,0 @@
-export interface Notification {
-    id: number
-    label: string
-}
