+++ src/entities/webtoon-search/model/types.ts
